feat: Add Alias for Reusing Previous Failed Commit
Adds the `fix-commit` alias, which reopens the previous failed commit's
message for editing and reuse rather than requiring retype from scratch.
The message file is located with the help of the shell rather than
assuming it always lives at the top-level `.git` directory, since
that assumption breaks for worktrees, where the top-directory `.git` is
a file instead of the normal metadata directory.
Extra flags and arguments are passed into the command with the use of
`$@`, intentionally left (slightly) unprotected to favour flexibility.

Note that this alias uses the `sh` command under the hood to
enable the use of command substitution.
As such, `sh` must be available in `$PATH` when this alias is invoked
and be a shell conformant to POSIX standards.
For Unix-based platforms this will mostly not be a problem as
`sh` is most likely to be the system Bash shell.
For Windows platforms, the default distribution "Git for Windows"
comes with a bundled and executable Bash shell, but this must be
available in the shell on which Git is used and this alias is invoked.

Signed-off-by: Paul Kim <44695374+thekpaul@users.noreply.github.com>